fix(flow-deletion-issue): moved the current flow context to the list
